web-app/backend/pyproject.toml
[project]
name = "portfolio-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "gspread>=6.0.0",
    "google-auth>=2.30.0",
    "pydantic-settings>=2.0.0",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.8.0",
    "ty>=0.0.1a7",
]

[tool.ruff.lint]
select = ["E", "F", "I"]
